コンポーネント化
