+++ app/kamokusu/kamokusu.tsx
@@ -1,0 +1,37 @@
+"use client";
+
+import { useState } from 'react';
+import "./kamokusu.css";
+
+export const SubjectInput=()=> {
+  const [subjectCount, setSubjectCount] = useState<string>("");
+
+  const handleSubmit = (e: React.FormEvent) => {
+    e.preventDefault();
+    alert(`科目数: ${subjectCount}`);
+  };
+
+  return (
+    <div className="container">
+      <h1 className="title">GPA計算機</h1>
+      <form onSubmit={handleSubmit}>
+        <label htmlFor="count" className="label">
+          科目数を入力してください
+        </label>
+        <input
+          id="count"
+          type="number"
+          min="1"
+          value={subjectCount}
+          onChange={(e) => setSubjectCount(e.target.value)}
+          className="input"
+          placeholder="例：5"
+          required
+        />
+        <button type="submit" className="button">
+          次へ
+        </button>
+      </form>
+    </div>
+  );
+}
